Add permission handling to dashboard
