+++ app/src/main/res/drawable/ic_bt_location_disabled.xml
@@ -1,0 +1,24 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="130dp"
+    android:height="130dp"
+    android:viewportWidth="130"
+    android:viewportHeight="130">
+  <path
+      android:pathData="M65.026,123.292C97.22,123.292 123.318,97.194 123.318,65C123.318,32.806 97.22,6.708 65.026,6.708C32.832,6.708 6.734,32.806 6.734,65C6.734,97.194 32.832,123.292 65.026,123.292Z"
+      android:fillColor="#BFE3F8"/>
+  <path
+      android:pathData="M65,0C29.094,0 0,29.094 0,65C0,100.906 29.094,130 65,130C100.906,130 130,100.906 130,65C130,29.094 100.906,0 65,0ZM65,123.292C32.812,123.292 6.708,97.188 6.708,65C6.708,32.812 32.812,6.708 65,6.708C97.188,6.708 123.292,32.812 123.292,65C123.292,97.188 97.188,123.292 65,123.292Z"
+      android:fillColor="#DFF2FC"/>
+  <path
+      android:pathData="M65,38.667C62.237,38.667 60,41.054 60,44C60,46.947 62.237,49.333 65,49.333C67.762,49.333 70,46.947 70,44C70,41.054 67.762,38.667 65,38.667ZM76.175,42.667C75.6,37.107 71.463,32.694 66.25,32.08V29.333H63.75V32.08C58.537,32.694 54.4,37.107 53.825,42.667H51.25V45.333H53.825C54.4,50.894 58.537,55.307 63.75,55.92V58.667H66.25V55.92C71.463,55.307 75.6,50.894 76.175,45.333H78.75V42.667H76.175ZM65,53.333C60.162,53.333 56.25,49.16 56.25,44C56.25,38.84 60.162,34.667 65,34.667C69.838,34.667 73.75,38.84 73.75,44C73.75,49.16 69.838,53.333 65,53.333Z"
+      android:fillColor="#4979BC"/>
+  <path
+      android:pathData="M93.606,87.094C95.036,87.094 96.206,85.924 96.206,84.494V74.068C96.206,72.638 95.036,71.468 93.606,71.468C92.176,71.468 91.006,72.638 91.006,74.068V84.52C91.006,85.924 92.176,87.094 93.606,87.094Z"
+      android:fillColor="#D81D49"/>
+  <path
+      android:pathData="M102.29,78.332C101.198,77.344 99.612,77.448 98.65,78.436C97.688,79.424 97.61,81.088 98.832,82.206C100.756,83.948 101.744,86.6 101.276,89.278C100.652,92.918 97.48,95.778 93.606,95.778C89.628,95.778 86.586,92.866 85.962,89.408C85.468,86.704 86.378,84.026 88.38,82.206C89.602,81.114 89.524,79.424 88.562,78.436C87.6,77.448 86.014,77.344 84.922,78.358C81.88,81.114 80.294,85.196 80.736,89.33C81.438,95.908 87.002,100.978 93.658,100.978C100.236,100.978 105.644,95.908 106.476,89.772C107.022,85.482 105.462,81.218 102.29,78.332Z"
+      android:fillColor="#D81D49"/>
+  <path
+      android:pathData="M75.443,74.462L65.165,65.399C64.678,64.98 64.002,64.882 63.434,65.148C62.853,65.427 62.488,66.015 62.488,66.672V78.98L55.699,72.882C55.023,72.28 53.995,72.35 53.4,73.05C52.805,73.749 52.886,74.812 53.562,75.427L62.488,83.469V83.511L53.562,91.553C52.886,92.168 52.818,93.231 53.4,93.93C53.981,94.63 55.023,94.7 55.699,94.098L62.488,87.986V100.322C62.488,100.979 62.853,101.566 63.421,101.846C63.637,101.958 63.881,102 64.11,102C64.476,102 64.854,101.874 65.138,101.608L69.885,97.553C70.575,96.965 70.683,95.902 70.101,95.189C69.533,94.476 68.506,94.378 67.816,94.965L65.733,96.741V86.182L68.844,88.728C69.141,88.965 69.493,89.077 69.844,89.077C70.331,89.077 70.805,88.854 71.129,88.434C71.684,87.707 71.562,86.644 70.845,86.07L67.64,83.483L75.389,77.05C75.768,76.742 75.998,76.266 75.998,75.763C76.025,75.273 75.822,74.798 75.443,74.462ZM65.747,80.756V70.336L71.846,75.721L65.747,80.756Z"
+      android:fillColor="#4979BC"/>
+</vector>
