+++ app/src/main/res/drawable/ic_location_disabled.xml
@@ -1,0 +1,21 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="130dp"
+    android:height="130dp"
+    android:viewportWidth="130"
+    android:viewportHeight="130">
+  <path
+      android:pathData="M65.026,123.292C97.22,123.292 123.318,97.194 123.318,65C123.318,32.806 97.22,6.708 65.026,6.708C32.832,6.708 6.734,32.806 6.734,65C6.734,97.194 32.832,123.292 65.026,123.292Z"
+      android:fillColor="#BFE3F8"/>
+  <path
+      android:pathData="M65,0C29.094,0 0,29.094 0,65C0,100.906 29.094,130 65,130C100.906,130 130,100.906 130,65C130,29.094 100.906,0 65,0ZM65,123.292C32.812,123.292 6.708,97.188 6.708,65C6.708,32.812 32.812,6.708 65,6.708C97.188,6.708 123.292,32.812 123.292,65C123.292,97.188 97.188,123.292 65,123.292Z"
+      android:fillColor="#DFF2FC"/>
+  <path
+      android:pathData="M57.167,49.292C50.905,49.292 45.833,54.736 45.833,61.458C45.833,68.18 50.905,73.625 57.167,73.625C63.428,73.625 68.5,68.18 68.5,61.458C68.5,54.736 63.428,49.292 57.167,49.292ZM82.497,58.417C81.193,45.733 71.815,35.665 60,34.266V28H54.333V34.266C42.518,35.665 33.14,45.733 31.837,58.417H26V64.5H31.837C33.14,77.184 42.518,87.252 54.333,88.651V94.917H60V88.651C71.815,87.252 81.193,77.184 82.497,64.5H88.333V58.417H82.497ZM57.167,82.75C46.202,82.75 37.333,73.23 37.333,61.458C37.333,49.687 46.202,40.167 57.167,40.167C68.132,40.167 77,49.687 77,61.458C77,73.23 68.132,82.75 57.167,82.75Z"
+      android:fillColor="#4979BC"/>
+  <path
+      android:pathData="M93.606,87.094C95.036,87.094 96.206,85.924 96.206,84.494V74.068C96.206,72.638 95.036,71.468 93.606,71.468C92.176,71.468 91.006,72.638 91.006,74.068V84.52C91.006,85.924 92.176,87.094 93.606,87.094Z"
+      android:fillColor="#D81D49"/>
+  <path
+      android:pathData="M102.29,78.332C101.198,77.344 99.612,77.448 98.65,78.436C97.688,79.424 97.61,81.088 98.832,82.206C100.756,83.948 101.744,86.6 101.276,89.278C100.652,92.918 97.48,95.778 93.606,95.778C89.628,95.778 86.586,92.866 85.962,89.408C85.468,86.704 86.378,84.026 88.38,82.206C89.602,81.114 89.524,79.424 88.562,78.436C87.6,77.448 86.014,77.344 84.922,78.358C81.88,81.114 80.294,85.196 80.736,89.33C81.438,95.908 87.002,100.978 93.658,100.978C100.236,100.978 105.644,95.908 106.476,89.772C107.022,85.482 105.462,81.218 102.29,78.332Z"
+      android:fillColor="#D81D49"/>
+</vector>
